+++ packages/ui/src/components/HistoryCommandPalette.tsx
@@ -8,25 +8,36 @@ import {
     CommandItem,
     CommandSeparator,
 } from "@/components/ui/command";
-import { Skeleton } from "@/components/ui/skeleton";
 import { cn } from "@/lib/utils";
 import { formatPathTail } from "@/lib/path";
-import { MessageSquare, Play, Loader2, Clock } from "lucide-react";
-import type { ResumeSessionOption } from "@/lib/types";
+import { FolderOpen, Play, Loader2 } from "lucide-react";
+
+export interface HistoricalSession {
+    sessionId: string;
+    cwd: string;
+    shareUrl: string;
+    startedAt: string;
+    lastActiveAt: string;
+    endedAt: string | null;
+    isEphemeral: boolean;
+    expiresAt: string | null;
+    isPinned: boolean;
+    runnerId: string | null;
+    runnerName: string | null;
+    sessionName: string | null;
+}
 
 export interface HistoryCommandPaletteProps {
     open: boolean;
     onOpenChange: (open: boolean) => void;
-    /** Resume sessions from the runner (same source as /resume) */
-    sessions: ResumeSessionOption[];
-    /** Whether the session list is currently loading */
-    loading: boolean;
-    /** Trigger a refresh of the resume sessions list */
-    onRefresh: () => void;
-    /** Called when user selects a session to view */
+    /** Set of session IDs currently live (filtered out of history) */
+    liveSessionIds: Set<string>;
+    /** Set of runner IDs that are currently online */
+    onlineRunnerIds: Set<string>;
+    /** Called when user selects a historical session to view its snapshot */
     onOpenSession: (sessionId: string) => void;
-    /** Called when user wants to resume a session */
-    onResumeSession?: (sessionId: string) => void;
+    /** Called when user wants to resume a session on a runner */
+    onResumeSession?: (sessionId: string, runnerId: string, cwd: string) => void;
 }
 
 function formatRelativeDate(isoString: string): string {
@@ -60,12 +71,12 @@ function getDateGroup(isoString: string): string {
     return date.toLocaleDateString(undefined, { month: "long", year: "numeric" });
 }
 
-function groupByDate(sessions: ResumeSessionOption[]): { label: string; sessions: ResumeSessionOption[] }[] {
-    const groups: Map<string, ResumeSessionOption[]> = new Map();
+function groupByDate(sessions: HistoricalSession[]): { label: string; sessions: HistoricalSession[] }[] {
+    const groups: Map<string, HistoricalSession[]> = new Map();
     const order: string[] = [];
 
     for (const s of sessions) {
-        const label = getDateGroup(s.modified);
+        const label = getDateGroup(s.lastActiveAt);
         if (!groups.has(label)) {
             groups.set(label, []);
             order.push(label);
@@ -76,62 +87,96 @@ function groupByDate(sessions: ResumeSessionOption[]): { label: string; sessions
     return order.map((label) => ({ label, sessions: groups.get(label)! }));
 }
 
-/* ── Skeleton rows shown while loading ──────────────────────────────────── */
-
-function SkeletonRows() {
-    return (
-        <div className="p-2 space-y-1" role="status" aria-label="Loading sessions">
-            {/* Fake group heading */}
-            <Skeleton className="h-3 w-12 ml-2 mb-2 mt-1" />
-            {[1, 2, 3, 4, 5].map((i) => (
-                <div key={i} className="flex items-center gap-2.5 px-2 py-2.5">
-                    <Skeleton className="h-8 w-8 rounded-md flex-shrink-0" />
-                    <div className="flex-1 min-w-0 space-y-1.5">
-                        <div className="flex items-center justify-between gap-2">
-                            <Skeleton className="h-3.5 rounded" style={{ width: `${40 + (i * 13) % 35}%` }} />
-                            <Skeleton className="h-2.5 w-10 rounded flex-shrink-0" />
-                        </div>
-                        <Skeleton className="h-2.5 rounded" style={{ width: `${55 + (i * 17) % 30}%` }} />
-                    </div>
-                </div>
-            ))}
-        </div>
-    );
-}
+const PAGE_SIZE = 50;
 
 export const HistoryCommandPalette = React.memo(function HistoryCommandPalette({
     open,
     onOpenChange,
-    sessions,
-    loading,
-    onRefresh,
+    liveSessionIds,
+    onlineRunnerIds,
     onOpenSession,
     onResumeSession,
 }: HistoryCommandPaletteProps) {
+    const [sessions, setSessions] = React.useState<HistoricalSession[]>([]);
+    const [loading, setLoading] = React.useState(false);
+    const [initialLoaded, setInitialLoaded] = React.useState(false);
+    const [hasMore, setHasMore] = React.useState(false);
     const [resumingSessionId, setResumingSessionId] = React.useState<string | null>(null);
 
-    // Refresh when the palette opens, but skip if data was fetched recently
-    const lastOpenRef = React.useRef(false);
-    const lastFetchRef = React.useRef(0);
-    React.useEffect(() => {
-        if (open && !lastOpenRef.current) {
-            const now = Date.now();
-            if (now - lastFetchRef.current > 5_000 || sessions.length === 0) {
-                lastFetchRef.current = now;
-                onRefresh();
+    const fetchSessions = React.useCallback(async (offset: number = 0) => {
+        setLoading(true);
+        try {
+            const params = new URLSearchParams({ limit: String(PAGE_SIZE + 1) });
+            if (offset > 0) params.set("offset", String(offset));
+            const res = await fetch(`/api/sessions?${params.toString()}`, { credentials: "include" });
+            if (!res.ok) return;
+            const body = await res.json();
+            const persisted: HistoricalSession[] = Array.isArray(body?.persistedSessions)
+                ? body.persistedSessions.map((s: any) => ({
+                    sessionId: s.sessionId,
+                    cwd: s.cwd ?? "",
+                    shareUrl: s.shareUrl ?? "",
+                    startedAt: s.startedAt ?? "",
+                    lastActiveAt: s.lastActiveAt ?? "",
+                    endedAt: s.endedAt ?? null,
+                    isEphemeral: s.isEphemeral ?? false,
+                    expiresAt: s.expiresAt ?? null,
+                    isPinned: s.isPinned ?? false,
+                    runnerId: s.runnerId ?? null,
+                    runnerName: s.runnerName ?? null,
+                    sessionName: s.sessionName ?? null,
+                }))
+                : [];
+
+            const hasMoreItems = persisted.length > PAGE_SIZE;
+            const pageItems = hasMoreItems ? persisted.slice(0, PAGE_SIZE) : persisted;
+
+            if (offset > 0) {
+                setSessions((prev) => {
+                    const existingIds = new Set(prev.map((s) => s.sessionId));
+                    const newItems = pageItems.filter((s) => !existingIds.has(s.sessionId));
+                    return [...prev, ...newItems];
+                });
+            } else {
+                setSessions(pageItems);
             }
+            setHasMore(hasMoreItems);
+            setInitialLoaded(true);
+        } catch {
+            // best-effort
+        } finally {
+            setLoading(false);
+        }
+    }, []);
+
+    // Fetch when the palette opens
+    React.useEffect(() => {
+        if (open && !initialLoaded) {
+            fetchSessions();
+        }
+    }, [open, initialLoaded, fetchSessions]);
+
+    // Refresh on re-open (stale data check)
+    const lastOpenRef = React.useRef(false);
+    React.useEffect(() => {
+        if (open && !lastOpenRef.current && initialLoaded) {
+            // Re-opening — refresh the first page
+            fetchSessions(0);
         }
         lastOpenRef.current = open;
-    }, [open, onRefresh, sessions.length]);
+    }, [open, initialLoaded, fetchSessions]);
 
-    const handleResume = React.useCallback((e: React.MouseEvent, s: ResumeSessionOption) => {
+    const handleResume = React.useCallback(async (e: React.MouseEvent, s: HistoricalSession) => {
         e.stopPropagation();
         e.preventDefault();
-        if (!onResumeSession) return;
-        setResumingSessionId(s.id);
-        onResumeSession(s.id);
-        onOpenChange(false);
-        setTimeout(() => setResumingSessionId(null), 2000);
+        if (!s.runnerId || !onResumeSession) return;
+        setResumingSessionId(s.sessionId);
+        try {
+            await onResumeSession(s.sessionId, s.runnerId, s.cwd);
+            onOpenChange(false);
+        } finally {
+            setResumingSessionId(null);
+        }
     }, [onResumeSession, onOpenChange]);
 
     const handleSelect = React.useCallback((sessionId: string) => {
@@ -139,9 +184,13 @@ export const HistoryCommandPalette = React.memo(function HistoryCommandPalette({
         onOpenChange(false);
     }, [onOpenSession, onOpenChange]);
 
-    const dateGroups = React.useMemo(() => groupByDate(sessions), [sessions]);
+    // Filter out live sessions
+    const filteredSessions = React.useMemo(
+        () => sessions.filter((s) => !liveSessionIds.has(s.sessionId)),
+        [sessions, liveSessionIds],
+    );
 
-    const showSkeleton = loading && sessions.length === 0;
+    const dateGroups = React.useMemo(() => groupByDate(filteredSessions), [filteredSessions]);
 
     return (
         <CommandDialog
@@ -149,102 +198,115 @@ export const HistoryCommandPalette = React.memo(function HistoryCommandPalette({
             onOpenChange={onOpenChange}
             title="Session History"
             description="Search and resume past sessions"
-            className="max-w-lg max-md:max-w-[calc(100%-1rem)]"
+            className="max-w-lg"
             showCloseButton={false}
         >
-            <CommandInput placeholder="Search sessions…" />
-            <CommandList className="max-h-[min(70vh,420px)] md:max-h-[min(60vh,400px)]">
-                {showSkeleton ? (
-                    <SkeletonRows />
-                ) : (
-                    <>
-                        <CommandEmpty>
-                            <div className="flex flex-col items-center gap-2 py-4 text-muted-foreground">
-                                <Clock className="h-8 w-8 opacity-30" />
-                                <span className="text-sm">No sessions found</span>
-                            </div>
-                        </CommandEmpty>
+            <CommandInput placeholder="Search sessions by name, path, or runner…" />
+            <CommandList className="max-h-[min(60vh,400px)]">
+                <CommandEmpty>
+                    {loading ? (
+                        <div className="flex items-center justify-center gap-2">
+                            <Loader2 className="h-4 w-4 animate-spin" />
+                            <span>Loading sessions…</span>
+                        </div>
+                    ) : (
+                        "No sessions found"
+                    )}
+                </CommandEmpty>
 
-                        {dateGroups.map((group, gi) => (
-                            <React.Fragment key={group.label}>
-                                {gi > 0 && <CommandSeparator />}
-                                <CommandGroup heading={group.label}>
-                                    {group.sessions.map((s) => {
-                                        const isResuming = resumingSessionId === s.id;
-                                        const displayName = s.name?.trim() || `Session ${s.id.slice(0, 8)}…`;
-                                        const preview = s.firstMessage && s.firstMessage !== "(no messages)"
-                                            ? s.firstMessage
-                                            : null;
+                {dateGroups.map((group, gi) => (
+                    <React.Fragment key={group.label}>
+                        {gi > 0 && <CommandSeparator />}
+                        <CommandGroup heading={group.label}>
+                            {group.sessions.map((s) => {
+                                const canResume = !!s.runnerId && onlineRunnerIds.has(s.runnerId) && !!onResumeSession;
+                                const isResuming = resumingSessionId === s.sessionId;
+                                const displayName = s.sessionName?.trim() || `Session ${s.sessionId.slice(0, 8)}…`;
 
-                                        const keywords = [
-                                            s.id,
-                                            s.path,
-                                            s.name ?? "",
-                                            s.firstMessage ?? "",
-                                        ].filter(Boolean);
+                                // Build search keywords for cmdk filtering
+                                const keywords = [
+                                    s.sessionId,
+                                    s.cwd,
+                                    s.runnerName ?? "",
+                                    s.runnerId ?? "",
+                                    s.sessionName ?? "",
+                                ].filter(Boolean);
 
-                                        return (
-                                            <CommandItem
-                                                key={s.id}
-                                                value={`${displayName} ${s.path}`}
-                                                keywords={keywords}
-                                                onSelect={() => handleSelect(s.id)}
-                                                className="flex items-center gap-2.5 py-3 md:py-2.5 rounded-md"
-                                            >
-                                                <div className="flex-shrink-0 flex items-center justify-center w-8 h-8 md:w-7 md:h-7 rounded-md bg-muted/60">
-                                                    <MessageSquare className="size-4 md:size-3.5 text-muted-foreground/70" />
-                                                </div>
-                                                <div className="flex-1 min-w-0">
-                                                    <div className="flex items-baseline justify-between gap-2 min-w-0">
-                                                        <span className="truncate text-sm font-medium leading-tight">
-                                                            {displayName}
-                                                        </span>
-                                                        <span className="text-[0.6rem] text-muted-foreground/60 flex-shrink-0 whitespace-nowrap tabular-nums">
-                                                            {formatRelativeDate(s.modified)}
-                                                        </span>
-                                                    </div>
-                                                    {preview && (
-                                                        <p className="mt-0.5 text-xs text-muted-foreground/50 truncate leading-tight">
-                                                            {preview.length > 72 ? `${preview.slice(0, 72)}…` : preview}
-                                                        </p>
-                                                    )}
-                                                    <div className="flex items-center gap-1 mt-0.5 min-w-0">
-                                                        <span className="text-[0.65rem] text-muted-foreground/40 truncate leading-tight" title={s.path}>
-                                                            {formatPathTail(s.path, 2)}
-                                                        </span>
-                                                    </div>
-                                                </div>
-
-                                                {/* Resume — icon-only on mobile, label on desktop */}
-                                                {onResumeSession && (
-                                                    <button
-                                                        type="button"
-                                                        onClick={(e) => handleResume(e, s)}
-                                                        disabled={isResuming}
-                                                        className={cn(
-                                                            "flex-shrink-0 flex items-center justify-center gap-1 rounded-md transition-colors",
-                                                            "h-8 w-8 md:h-auto md:w-auto md:px-2 md:py-1",
-                                                            isResuming
-                                                                ? "text-muted-foreground/40"
-                                                                : "text-green-600 dark:text-green-400 hover:bg-green-500/10 active:bg-green-500/20",
-                                                        )}
-                                                        title="Resume session"
-                                                        aria-label="Resume session"
-                                                    >
-                                                        {isResuming ? (
-                                                            <Loader2 className="h-3.5 w-3.5 animate-spin" />
-                                                        ) : (
-                                                            <Play className="h-3.5 w-3.5" />
-                                                        )}
-                                                        <span className="hidden md:inline text-xs font-medium">Resume</span>
-                                                    </button>
+                                return (
+                                    <CommandItem
+                                        key={s.sessionId}
+                                        value={`${displayName} ${s.cwd} ${s.runnerName ?? ""}`}
+                                        keywords={keywords}
+                                        onSelect={() => handleSelect(s.sessionId)}
+                                        className="flex items-center gap-2.5 py-2.5"
+                                    >
+                                        <div className="flex-shrink-0 flex items-center justify-center w-7 h-7 rounded-md bg-muted/50">
+                                            <FolderOpen className="size-3.5 text-muted-foreground" />
+                                        </div>
+                                        <div className="flex-1 min-w-0">
+                                            <div className="flex items-baseline justify-between gap-2 min-w-0">
+                                                <span className="truncate text-sm font-medium">
+                                                    {displayName}
+                                                </span>
+                                                <span className="text-[0.65rem] text-muted-foreground flex-shrink-0 whitespace-nowrap">
+                                                    {formatRelativeDate(s.lastActiveAt)}
+                                                </span>
+                                            </div>
+                                            <div className="flex items-center gap-1 mt-0.5 min-w-0">
+                                                <span className="text-xs text-muted-foreground/70 truncate" title={s.cwd}>
+                                                    {formatPathTail(s.cwd, 2)}
+                                                </span>
+                                                {(s.runnerName || s.runnerId) && (
+                                                    <span className="text-[0.65rem] text-muted-foreground/50 truncate max-w-[6rem]">
+                                                        · {s.runnerName || `Runner ${s.runnerId?.slice(0, 8)}…`}
+                                                    </span>
                                                 )}
-                                            </CommandItem>
-                                        );
-                                    })}
-                                </CommandGroup>
-                            </React.Fragment>
-                        ))}
+                                            </div>
+                                        </div>
+
+                                        {/* Resume button */}
+                                        {canResume && (
+                                            <button
+                                                type="button"
+                                                onClick={(e) => handleResume(e, s)}
+                                                disabled={isResuming}
+                                                className={cn(
+                                                    "flex-shrink-0 flex items-center gap-1 px-2 py-1 rounded-md text-xs font-medium transition-colors",
+                                                    isResuming
+                                                        ? "text-muted-foreground/50"
+                                                        : "text-green-600 dark:text-green-400 hover:bg-green-500/10",
+                                                )}
+                                                title={`Resume on ${s.runnerName || "runner"}`}
+                                            >
+                                                {isResuming ? (
+                                                    <Loader2 className="h-3 w-3 animate-spin" />
+                                                ) : (
+                                                    <Play className="h-3 w-3" />
+                                                )}
+                                                <span>Resume</span>
+                                            </button>
+                                        )}
+                                    </CommandItem>
+                                );
+                            })}
+                        </CommandGroup>
+                    </React.Fragment>
+                ))}
+
+                {/* Load more */}
+                {hasMore && (
+                    <>
+                        <CommandSeparator />
+                        <div className="p-2">
+                            <button
+                                type="button"
+                                onClick={() => fetchSessions(sessions.length)}
+                                disabled={loading}
+                                className="w-full px-3 py-2 text-xs font-medium text-muted-foreground hover:text-foreground hover:bg-accent rounded-md transition-colors disabled:opacity-40 disabled:cursor-not-allowed"
+                            >
+                                {loading ? "Loading…" : "Load more sessions"}
+                            </button>
+                        </div>
                     </>
                 )}
             </CommandList>
